+++ core-ng/src/main/java/core/framework/http/ContentType.java
@@ -21,6 +21,8 @@ public final class ContentType {
     public static final ContentType TEXT_XML = create("text/xml", UTF_8);
     public static final ContentType APPLICATION_JSON = create("application/json", UTF_8);
     public static final ContentType APPLICATION_JAVASCRIPT = create("application/javascript", UTF_8);
+    // form body content type doesn't use charset normally, refer to https://www.w3.org/TR/html5/sec-forms.html#urlencoded-form-data
+    public static final ContentType APPLICATION_FORM_URLENCODED = ContentType.create("application/x-www-form-urlencoded", null);
     public static final ContentType APPLICATION_OCTET_STREAM = create("application/octet-stream", null);
     public static final ContentType IMAGE_PNG = create("image/png", null);
 
@@ -67,9 +69,9 @@ public final class ContentType {
         return new ContentType(contentType, mediaType, charset);
     }
 
-    private final Charset charset;
     private final String contentType;
     private final String mediaType;
+    private final Charset charset;
 
     private ContentType(String contentType, String mediaType, Charset charset) {
         this.contentType = contentType;