+++ core-ng/src/main/java/core/framework/http/HTTPRequest.java
@@ -13,51 +13,25 @@ import static java.nio.charset.StandardCharsets.UTF_8;
  * @author neo
  */
 public final class HTTPRequest {
-    // due to history issue, many web server may have issue if send charset=utf-8 explicitly, so not use charset
-    private static final ContentType APPLICATION_FORM_URLENCODED = ContentType.create("application/x-www-form-urlencoded", null);
+    public final String uri;
+    public final HTTPMethod method;
+    public final Map<String, String> params = Maps.newLinkedHashMap();
+    public final Map<String, String> headers = Maps.newLinkedHashMap();    // make headers/params order deterministic
 
-    private final String uri;
-    private final HTTPMethod method;
-    private final Map<String, String> headers = Maps.newLinkedHashMap();    // make headers/params order deterministic, (e.g. for use cases where http request needs to be signed by hash)
-    private final Map<String, String> params = Maps.newLinkedHashMap();
-    private ContentType contentType;
     private byte[] body;
+    private ContentType contentType;
 
     public HTTPRequest(HTTPMethod method, String uri) {
         this.method = method;
         this.uri = uri;
     }
 
-    public String uri() {
-        return uri;
-    }
-
-    public HTTPMethod method() {
-        return method;
-    }
-
-    public Map<String, String> headers() {
-        return headers;
-    }
-
-    public void header(String name, String value) {
-        headers.put(name, value);
-    }
-
     public void accept(ContentType contentType) {
-        header(HTTPHeaders.ACCEPT, contentType.toString());
+        headers.put(HTTPHeaders.ACCEPT, contentType.toString());
     }
 
     public void basicAuth(String user, String password) {
-        header(HTTPHeaders.AUTHORIZATION, "Basic " + Encodings.base64(user + ':' + password));
-    }
-
-    public Map<String, String> params() {
-        return params;
-    }
-
-    public void param(String name, String value) {
-        params.put(name, value);
+        headers.put(HTTPHeaders.AUTHORIZATION, "Basic " + Encodings.base64(user + ':' + password));
     }
 
     public byte[] body() {
@@ -72,12 +46,13 @@ public final class HTTPRequest {
     public void body(byte[] body, ContentType contentType) {
         this.body = body;
         this.contentType = contentType;
+        headers.put(HTTPHeaders.CONTENT_TYPE, contentType.toString());
     }
 
     public void form(Map<String, String> form) {
         var builder = new StringBuilder(256);
         HTTPRequestHelper.urlEncoding(builder, form);
-        body(Strings.bytes(builder.toString()), APPLICATION_FORM_URLENCODED);
+        body(Strings.bytes(builder.toString()), ContentType.APPLICATION_FORM_URLENCODED);
     }
 
     public ContentType contentType() {